second commit

+++ BL/Interfaces/ICostumerService.cs
@@ -11,7 +11,7 @@ namespace BL.Interfaces
     public interface ICostumerService
     {
         public Task<List<Costumer>> GetAllCostumers();
-        public Task<Costumer> GetCostumerById(long id);
+        public Task<Costumer> GetCostumerById(int id);
         public Task<bool> AddCostumer(CostumerDto costumer);
         //public Task<bool> AddHoursDonation(int hours, long id);
         //public Task<bool> RemoveHoursAvailable(int hours, long id);
