+++ DAL/Interfaces/ICostumerData.cs
@@ -11,9 +11,7 @@ namespace DAL.Interfaces
     public interface ICostumerData
     {
         public Task<List<Costumer>> GetAllCostumers();
-        public Task<Costumer> GetCostumerById(long id);
+        public Task<Costumer> GetCostumerById(int id);
         public Task<bool> AddCostumer(CostumerDto user);
-        //public Task<bool> AddHoursDonation(int hours, long id);
-        //public Task<bool> RemoveHoursAvailable(int hours, long id);
     }
 }
